extractor_api_app/services/image_utils.py
import cv2
import numpy as np
from PIL import Image

def preprocess_image(image):
    """Preprocess images to improve OCR accuracy."""
    img = np.array(image)
    gray = cv2.cvtColor(img, cv2.COLOR_RGB2GRAY)  # Convert to grayscale
    _, thresh = cv2.threshold(gray, 150, 255, cv2.THRESH_BINARY + cv2.THRESH_OTSU)
    return Image.fromarray(thresh)
