extractor_api_app/requirements.txt
fastapi
uvicorn
pdf2image
pytesseract
pillow
opencv-python
numpy
python-multipart


#sudo apt install tesseract-ocr
